Let's be more accurate with the type conversion.
git-svn-id: svn+ssh://gaia.geo.uni-mainz.de/local/home/lkausb/svn/LaMEM/trunk@6048 26174531-3d5a-4d31-aba0-7a6658d032cb

+++ src/fdstag/paraViewOutBin.c
@@ -126,7 +126,7 @@ void OutBufDump(OutBuf *outbuf)
 	int nbytes;
 
 	// compute number of bytes
-	nbytes = outbuf->cn*(int)sizeof(float);
+	nbytes = (int)outbuf->cn*(int)sizeof(float);
 
 	// dump number of bytes
 	fwrite(&nbytes, sizeof(int), 1, outbuf->fp);
